#5: add instructions to tear down site, don't re-install if site already exists

+++ .gitpod/install_joomla_beta.sh
@@ -25,6 +25,8 @@ echo "* Configuring site"
 joomla site:configure ${APACHE_DOCROOT_IN_REPO} --www=$GITPOD_REPO_ROOT --mysql-login=root:;
 
 #Ensure that we can upgrade insecure requests via the apache conf
-ln -s /etc/apache2/mods-available/headers.load /etc/apache2/mods-enabled/headers.load
+if [ ! -e "/etc/apache2/mods-enabled/headers.load" ]; then
+  ln -s /etc/apache2/mods-available/headers.load /etc/apache2/mods-enabled/headers.load
+fi;
 
 apachectl start
